:recycle: rename class name related to group
